Refactoring and create a new db session and helpers
