+++ {{cookiecutter.path_name}}/{{cookiecutter.project_slug}}/__init__.py
@@ -1,0 +1,5 @@
+import os
+import sys
+
+
+sys.path.insert(0, os.path.join(os.path.dirname(__file__)))
