+++ {{cookiecutter.path_name}}/{{cookiecutter.project_slug}}/apps/customers/api.py
@@ -4,6 +4,9 @@ from restless.preparers import FieldsPreparer
 
 from contrib.handlers import RestHandler
 
+from contrib.db import session
+from contrib.db.utils import get_or_404
+
 from .models import Customer
 
 
@@ -17,10 +20,10 @@ class CustomerHandler(RestHandler):
 
     @gen.coroutine
     def list(self):
-        customers = self.model.query.slice(0, 10)
+        customers = session.query(self.model).slice(0, 10)
         return customers
 
     @gen.coroutine
     def detail(self, id):
-        customer = self.model.get_or_404(id)
+        customer = get_or_404(self.model, id)
         return customer
