+++ {{cookiecutter.path_name}}/{{cookiecutter.project_slug}}/contrib/db/__init__.py
@@ -1,0 +1,2 @@
+from .session import session
+from .models import Model
