--- {{cookiecutter.path_name}}/{{cookiecutter.project_slug}}/contrib/shell.py
@@ -1,19 +1,0 @@
-#!/usr/bin/env python
-
-import os
-import readline
-from pprint import pprint
-
-
-def make_shell():
-    from contrib.server import *
-
-    os.environ['PYTHONINSPECT'] = 'True'
-
-
-def run():
-    print('Loading...')
-    make_shell()
-
-if __name__ == '__main__':
-    run()
