+++ {{cookiecutter.path_name}}/{{cookiecutter.project_slug}}/settings/development.py
@@ -1,1 +1,3 @@
 from .base import *
+
+SQL_ECHO = True
