+++ {{cookiecutter.path_name}}/{{cookiecutter.project_slug}}/settings/production.py
@@ -1,3 +1,1 @@
 from .base import *
-
-SQL_ECHO = False
