+++ {{cookiecutter.path_name}}/{{cookiecutter.project_slug}}/settings/test.py
@@ -3,8 +3,6 @@ import os
 from .base import * 
 
 
-SQL_ECHO = False 
-
 TEST_DB_PATH = os.path.join(PROJECT_DIR, 'test_db.db')
 
 DATABASES = {
